+++ nos/inet/socket.h
@@ -74,7 +74,7 @@ namespace std
         typedef std::size_t result_type;
         result_type operator()(argument_type const &s) const
         {
-            return std::hash<int>()(s.fd());
+            return std::hash<int64_t>()(s.fd());
         }
     };
 }
